Update main.yml
We need to use an OS version that both GitHub provides and is listed in the Cortex system requirements.

Looking at your docs, Ubuntu 22.04 is listed. GitHub provides this as ubuntu-22.04.

Let's update the image-scan job to use ubuntu-22.04. We will keep the dependency installation step, as it's highly likely those libraries are still required.